Fix incorrect queue timeout setting
It appears that the `queue-timeout` setting is mislabeled, and should actually be `cli-timeout`.  After setting a `queue-timeout` option of 300 seconds and deploying, we noticed that our CLI Lambda timeout was still 60 seconds, and our Default Visibility Timeout for our SQS instance was still 70 seconds.

After changing `queue-timeout` to `cli-timeout` as a random guess and deploying, our CLI Lambda timeout indeed updated to 300 seconds, and our SQS Default Visibility Timeout increased to the expected 310 seconds.