Update introduction.md to explain /vapor-ui and local environment (#11)
The setup documentation doesn't mention that the /vapor-ui URI is only intended for use in an AWS environment, deployed via `vapor deploy`.

This PR adds further documentation to explain that it's intended for an AWS environment, and will show a 404 page when used on a local environment.

This felt necessary as otherwise, when first installing Vapor UI, it feels like it hasn't worked, and time is spent debugging and researching why `/vapor-ui` is inaccessible. This note helps avoid that time, by noting that it won't be accessible until deployed to an AWS environment.